server/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.0
bcrypt==4.1.2
Pillow==10.1.0

